refactor: consolidate some funcs to utils
