+++ client/fe/webpack/moduleRules.js
@@ -9,7 +9,7 @@
 
 const MiniCssExtractPlugin = require('mini-css-extract-plugin')
 const config = require('../config/config').config
-const resolveDir = require('./resolveDir').resolveDir
+const resolveDir = require('./util').resolveDir
 
 /**
  * List of modules which need to be transpiled with Babel
